Refactor: Moved tweet engagement functions to xApiUtils.ts for better separation of concerns
Related to #1391
